use higher precision timing/dates

+++ core/Emailer.php
@@ -133,7 +133,7 @@ class Emailer extends StandardObject
     /**
      * Gets the config as a printable client object
      * @return array `{type:string, hosts:?string, username:?string, password:bool,
-         from_address:string, from_name:?string, features:{reply:bool}, dates:{created:int}}`
+         from_address:string, from_name:?string, features:{reply:bool}, dates:{created:float}}`
      */
     public function GetClientObject() : array
     {
